requirements.txt
streamlit==1.39.0
requests==2.32.3
pdfplumber==0.11.7
scikit-learn==1.5.2
faiss-cpu==1.8.0
numpy==1.26.4
joblib==1.4.2
tiktoken==0.7.0
python-dotenv==1.0.1
pandas==2.2.3
gtts==2.5.4
